requirements.txt
numpy~=2.1.2
loguru~=0.7.3
spiref~=1.0.4
torch~=2.7.0+cu118
transformers~=4.52.3
tqdm~=4.67.1
peft~=0.15.2
pyyaml~=6.0.2
huggingface-hub~=0.31.4
datasets~=3.6.0
requests~=2.32.3